% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dba_clust_associate.R
\name{dba_clust_associate}
\alias{dba_clust_associate}
\title{dba_clust_associate}
\usage{
dba_clust_associate(dt, pc.dba, explanatory_variable = colnames(dt),
  list_cluster_k, print_time = FALSE)
}
\arguments{
\item{dt}{A matrix, json or data frame where each row is a time series, or a list where each element is a time series}

\item{explanatory_variable}{list of covariates (optional),}

\item{list_cluster_k}{Number of desired clusters. It can be a numeric vector with different values.}

\item{print_time}{computational time}
}
\value{
An dataframe
}
\description{
Associate numerical times series to the closest centroid for each freatures with the \code{\link{dtw}} function.
Based o DTW distance.
}
\examples{
dba_clust_associate(dt = data.train[1:1000,listInteger],  list_cluster_k = list_cluster_k)
}
