% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dba_clust_generate.R
\name{dba_clust_generate}
\alias{dba_clust_generate}
\title{dba_clust_generate}
\usage{
dba_clust_generate(dt, type = "partitional",
  list_cluster_k = list_cluster_k, learn_size = (nrow(dt)),
  explanatory_variable = colnames(dt), print_time = FALSE,
  arm_for_learn = "all", print_graph = "all")
}
\arguments{
\item{dt}{A matrix, json or data frame where each row is a time series, or a list where each element is a time series}

\item{type}{What type of clustering method to use: "partitional", "hierarchical", "tadpole" or "fuzzy" (optional),}

\item{list_cluster_k}{Number of desired clusters. It can be a numeric vector with different values.}

\item{learn_size}{number of items dedicated to the learnset (step 1) ,}

\item{explanatory_variable}{list of covariates (optional),}

\item{print_time}{computational time}

\item{arm_for_learn}{arm dedicated to the learnset (step 1) (optional),}

\item{print_graph}{print clusters and controid c("all", "centroid","cluster","none")}
}
\value{
An object of class \code{\link{TSClusters-class}}
}
\description{
Generate a clustering of freatures with \code{\link{dba}} and \code{\link{tsclust}} functions.
Apply a global averaging method for time series under DTW (Petitjean, Ketterlin and Gancarski 2011) of numerical times series features.
}
\examples{
dba_clust_generate(dt = data.train[1:1000,listInteger],  list_cluster_k = list_cluster_k , print_graph = "none")
}
