% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_generate.R
\name{mean_generate}
\alias{mean_generate}
\title{mean_generate}
\usage{
mean_generate(dt, learn_size = (nrow(dt)),
  explanatory_variable = colnames(dt), print_time = FALSE)
}
\arguments{
\item{dt}{A matrix, json or data frame where each row is a time series, or a list where each element is a time series}

\item{explanatory_variable}{list of covariates (optional),}

\item{print_time}{computational time}

\item{print_graph}{print clusters and controid c("all", "centroid","cluster","none")}
}
\value{
An object of class \code{\link{TSClusters-class}}
}
\description{
Add global means for each time series
}
\examples{
mean_generate(dt = data.train)
}
